backend/pyproject.toml
[tool.poetry]
name = "peer-bonus-backend"
version = "0.1.0"
description = "Backend for Peer Bonus application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.25"
pydantic = {extras = ["email"], version = "^2.6.1"}
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
python-dateutil = "^2.8.2"
strawberry-graphql = {extras = ["fastapi"], version = "^0.217.1"}
pytest = "^7.4.4"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
ruff = "^0.2.2"
isort = "^5.13.2"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",  # Line too long - handled by black
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
show_error_context = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = ["sqlalchemy.*"]
ignore_missing_imports = true
